Develop

+++ src/Application/AppServices/OrderAppService.cs
@@ -1,8 +1,11 @@
-﻿using Application.Dtos.Orders;
+﻿using Application.Dtos;
+using Application.Dtos.Orders;
 using Application.Interfaces;
 using AutoMapper;
+using Domain.Entities.Orders;
 using Domain.Enums;
 using Domain.Exceptions;
+using Domain.Interfaces.ExternalServices;
 using Domain.Interfaces.Repositories.Orders;
 using Microsoft.Extensions.Logging;
 
@@ -10,29 +13,42 @@ namespace AppServices.Orders
 {
     public class OrderAppService : IOrderAppService
     {
-        private readonly ILogger<OrderAppService> _logger; 
+        private readonly ILogger<OrderAppService> _logger;
         private readonly IOrderRepository _orderRepository;
+        private readonly ICartAdapter _cartAdapter;
+        private readonly IPaymentAdapter _paymentAdapter;
+
         private readonly IMapper _mapper;
 
         public OrderAppService(
-            IOrderRepository orderRepository,  
+            IOrderRepository orderRepository,
             IMapper mapper,
+            ICartAdapter cartAdapter,
+            IPaymentAdapter paymentAdapter,
             ILogger<OrderAppService> logger)
         {
-            _orderRepository = orderRepository;  
+            _orderRepository = orderRepository;
+            _cartAdapter = cartAdapter;
+            _paymentAdapter = paymentAdapter;
             _mapper = mapper;
             _logger = logger;
         }
 
-        public async Task<OrderOutputDto?> GetOrderByIdAsync(Guid id)
+        public async Task<OrderOutputDto> GetOrderByIdAsync(Guid id)
         {
             _logger.LogInformation("Retrieving order with ID {OrderId}.", id);
+
+            if (id == Guid.Empty)
+            {
+                _logger.LogWarning("Invalid Order ID provided: {OrderId}.", id);
+                throw new ArgumentException("Order ID cannot be empty.", nameof(id));
+            }
 
             var order = await _orderRepository.GetOrderByIdAsync(id);
             if (order == null)
             {
-                _logger.LogInformation("Order with ID {OrderId} not found.", id);
-                throw new EntityNotFoundException(string.Format("Order with ID {0} not found.", id));
+                _logger.LogWarning("Order with ID {OrderId} not found.", id);
+                throw new EntityNotFoundException($"Order with ID {id} not found.");
             }
 
             _logger.LogInformation("Order with ID {OrderId} retrieved successfully.", id);
@@ -40,23 +56,56 @@ namespace AppServices.Orders
             return _mapper.Map<OrderOutputDto>(order);
         }
 
-        public Task<OrderOutputDto> CreateOrderByCartAsync(Guid cartId)
+        public async Task<OrderOutputDto> CreateOrderByCartAsync(CreateOrderRequestDto request)
         {
-            //TODO - RECEBER CARRINHO VIA REQUEST
+            if (request.CartId == Guid.Empty)
+            {
+                _logger.LogWarning("Invalid Cart ID provided: {CartId}.", request.CartId);
+                throw new InvalidCartIdException("Cart ID cannot be empty.");
+            }
 
-            //TODO - SALVAR NO BANCO DE ORDER
+            _logger.LogInformation("Creating order for cart with ID {CartId}...", request.CartId);
 
-            //TODO - ADICIONAR CHAMADA HTTP PARA DELETAR O CARRINHO 
+            var order = _mapper.Map<Order>(request);
 
-            throw new NotImplementedException();
+            var createdOrder = await _orderRepository.AddAsync(order);
+
+            _logger.LogInformation("Order created successfully with ID {OrderId}. Cart with ID {CartId} deleted after order creation.",
+                createdOrder.Id, request.CartId);
+
+            await _paymentAdapter.GeneratePayment(
+                orderId: createdOrder.Id,
+                model: new PaymentRequest
+                {
+                    OrderId = createdOrder.Id,
+                    TotalAmount = createdOrder.TotalAmount,
+                    Client = new PaymentRequest.ClientRequest
+                    {
+                        Name = "Teste",
+                        Cpf = "00000000019",
+                        Email = "teste@teste.com",
+                        Id = request.ClientId
+                    }
+                }
+            );
+
+            _ = _cartAdapter.DeleteCartByIdAsync(request.CartId);
+
+            return _mapper.Map<OrderOutputDto>(createdOrder);
         }
 
         public async Task DeleteOrderAsync(Guid orderId)
         {
             _logger.LogInformation("Attempting to delete order with ID: {OrderId}.", orderId);
 
+            if (orderId == Guid.Empty)
+            {
+                _logger.LogWarning("Invalid Order ID provided: {OrderId}.", orderId);
+                throw new ArgumentException("Order ID cannot be empty.", nameof(orderId));
+            }
+
             var existingOrder = await _orderRepository.GetByIdAsync(orderId)
-                ?? throw new EntityNotFoundException(string.Format("Order with ID {0} not found.", orderId));
+                ?? throw new EntityNotFoundException($"Order with ID {orderId} not found.");
 
             await _orderRepository.RemoveAsync(existingOrder);
 
@@ -65,41 +114,68 @@ namespace AppServices.Orders
 
         public async Task<List<OrderOutputDto>> GetOrderByStatusAsync(OrderStatus status)
         {
-            _logger.LogInformation("Retrieving order with status code {Status}.", status.ToString());
+            _logger.LogInformation("Retrieving orders with status {Status}.", status.ToString().Replace(Environment.NewLine, ""));
+
+            if (!Enum.IsDefined(typeof(OrderStatus), status))
+            {
+                _logger.LogWarning("Invalid order status provided: {Status}.", status.ToString().Replace(Environment.NewLine, ""));
+                throw new ArgumentException("Invalid order status.", nameof(status));
+            }
 
             var orders = await _orderRepository.GetOrderByStatusAsync(status);
             if (!orders.Any())
+            {
+                _logger.LogInformation("No orders found with status {Status}.", status.ToString().Replace(Environment.NewLine, ""));
                 return new List<OrderOutputDto>();
+            }
 
-            _logger.LogInformation("Order with status code {Status} retrieved successfully.", status.ToString());
+            _logger.LogInformation("Orders with status {Status} retrieved successfully.", status.ToString().Replace(Environment.NewLine, ""));
 
             return _mapper.Map<List<OrderOutputDto>>(orders);
         }
 
         public async Task<OrderOutputDto> MoveOrderToNextStatus(Guid id)
         {
+            if (id == Guid.Empty)
+            {
+                _logger.LogWarning("Invalid Order ID provided: {OrderId}.", id);
+                throw new ArgumentException("Order ID cannot be empty.", nameof(id));
+            }
+
             var originalOrder = await _orderRepository.GetByIdAsync(id)
-                ?? throw new EntityNotFoundException(string.Format("Order with ID {0} not found. Update aborted.", id));
+                ?? throw new EntityNotFoundException($"Order with ID {id} not found. Update aborted.");
 
-            //TODO - REMOVER E ADICIONAR CHAMADA AO SERVIÇO EXTERNO DE PAGAMENTOS PARA ATUALIZAR STATUS
-             
             originalOrder.SendToNextStatus();
-
             await _orderRepository.UpdateAsync(originalOrder);
 
-            return _mapper.Map<OrderOutputDto>(await _orderRepository.GetOrderByIdAsync(id));
+            _logger.LogInformation("Order with ID {OrderId} moved to next status successfully.", id);
+
+            var paymentStatusUpdate = await _paymentAdapter.MoveOrderToNextStatus(id);
+            if (!paymentStatusUpdate)
+            {
+                _logger.LogWarning("Failed to update payment status for order ID {OrderId}.", id);
+                throw new PaymentStatusUpdateException($"Payment status update failed for order ID {id}. Status transition incomplete.");
+            }
+
+            var updatedOrder = await _orderRepository.GetOrderByIdAsync(id);
+            _logger.LogInformation("Order with ID {OrderId} retrieved after status update.", id);
+
+            return _mapper.Map<OrderOutputDto>(updatedOrder);
         }
 
         public async Task<List<OrderOutputDto>> GetCurrrentOrders()
         {
-            _logger.LogInformation("Retrieving all orders.");
+            _logger.LogInformation("Retrieving all current orders.");
 
             var orders = await _orderRepository.GetCurrrentOrders();
 
             if (orders == null || !orders.Any())
+            {
+                _logger.LogInformation("No current orders found.");
                 return new List<OrderOutputDto>();
+            }
 
-            _logger.LogInformation("All orders retrieved successfully.");
+            _logger.LogInformation("All current orders retrieved successfully.");
 
             return _mapper.Map<List<OrderOutputDto>>(orders);
         }
